backend/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
boto3==1.34.0
python-dateutil==2.8.2
python-multipart==0.0.6